Receiver contract
This adds a receiver contract which automatically narrows when
inherited. Specifically, Receiver always refers to the class of the
thing receiving the method call.

    class Foo
      Contract Receiver => Receiver
      def some_method
        # ...
      end
    end

    class Bar < Foo; end
    # The inherited contract for #some_method will only pass when called
    # with a Bar (or a child of Bar). Passing in a Foo (or a child of
    # Foo that isn't also a child of Bar) would result in a
    # ContractError.

This contract didn't lend itself particularly well to the way Contracts
is structured. I had to add a receiver stack (implemented to be thread
and fiber safe), and perform arity checking to see if the receiver data
should be passed along.

My biggest complaint about this contract is that it doesn't print out
very nicely. Rather than `Expected: Receiver`, it would be nice to see
`Expected: #{name of the class receiver was checking for}`. Once again,
it would require a fair degree of refactoring the contract inspection
code to make that work well, so I'm leaving it alone at the moment.