Change syntax for no arguments
In many languages, a type signature for a function that increments a number by one looks like `Num -> Num`, which is represented fine in this library as `Contract Num => Num`. But for a function that just returns a value, say pi: What is pi? Pi is a Num (or Float/BigDecimal/whatever). Pi is not a `nil => Num`. I think it would improve the readability of the contract type signatures if there was a way to express the type of a declared value or computed value as a single type, and omit the `nil =>`.

This change allows you to write:

```rb
Contract Num
def pi
  3.14
end
```