Ignore Gemfile.lock
Having the lockfile actually prevented me from being able to `bundle install`. I had to do a `bundle update` instead (which changed the `Gemfile.lock`). Commiting a `Gemfile.lock` also increases the risk of library incompatibilities, since people installing the gem from RubyGems with `gem install contracts` or putting `gem 'contracts'` in their `Gemfile` will not respect the `Gemfile.lock`.

For more information on when and when not to commit a `Gemfile.lock`, see http://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/