Adds DescendantOf
This adds the ability to make assertions on the class hierarchy of a
class being passed as an argument. It should be useful when using
things like the Adapter Pattern or Dependency Injection.

Consider the following example:

```
class SlackClientAdapter < BaseClient
end

class HipchatClientAdapter < BaseClient
end

Contract C::DescendantOf[BaseClient] => BaseClient
def start_client(client_class)
  @client = client_class.new
end

start_client(SlackClientAdapter)
start_client(HipchatClientAdapter)
```

This is a demonstration of what I'm actually doing on a real app
(although the example itself is totally made up).

I want to be able to assert that the argument is not just a class, but a
descendant of `BaseClient` (this base client serves as an abstract class
this example. God, I feel like a Java developer all over again...)

I'm not sure if this is something that's already been proposed and
discarded, so I didn't take a lot of care in the organization of the
code I added.
Should this be accepted, I'll gladly make any required changes before
merging.