Fix StrictHash contract for extra keys (#254)
* Fix StrictHash contract for extra keys

`Contracts::StrictHash` didn't complain about extra entries in a given
hash. This was because of the missing assertion for keys. Specs hadn't
caught this case because `age` key had a wrong type anyway.

Compare keys for contract and a given hash and return false if they are
equal.

* Do not use a second lookup for the contract in StrictHash assertion

`all?` iterator already provides contract for the `valid?` check. Use
it instead of accessing `contract_hash` again.